requirements.txt
pdf2image>=1.16.0
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
google-genai>=0.2.0
numpy>=1.24.0
